requirements.txt
fastapi[standard]
beanie
motor
pydantic[email]
python-jose[cryptography]
bcrypt==4.0.1
passlib[bcrypt]
python-dotenv